+++ src/imio/smartweb/core/contents/rest/events/content.py
@@ -30,6 +30,13 @@ class IEventsView(model.Schema):
         title=_("Number of items to display"), default=20, required=True
     )
 
+    display_map = schema.Bool(
+        title=_("Display map"),
+        description=_("If selected, map will be displayed"),
+        required=False,
+        default=True,
+    )
+
 
 @implementer(IEventsView)
 class EventsView(Container):